refactor: RecentSpentGraph graph 이름을 지정하도록 수정
